+++ Source/FaaS.MVC/Views/React/DashBoard.cshtml
@@ -1,0 +1,9 @@
+﻿@{
+    ViewData["Title"] = "Dash Board";
+}
+
+<h2>Index</h2>
+
+<div id="react-root"></div>
+
+<script src="~/out/dashboard.out.js"></script>
