+++ Source/FaaS.MVC/Views/React/Login.cshtml
@@ -1,0 +1,9 @@
+﻿@{
+    ViewData["Title"] = "Login form";
+}
+
+<h2>Index</h2>
+
+<div id="react-root"></div>
+
+<script src="~/out/login.out.js"></script>
