+++ Source/FaaS.MVC/Views/Shared/_Layout.cshtml
@@ -32,6 +32,7 @@
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
                     <li><a asp-area="" asp-controller="Home" asp-action="Index">Home</a></li>
+                    <li><a asp-area="" asp-controller="Login" asp-action="Index">React Login</a></li>
                     <li class="dropdown-toggle" style="display:@(ViewData["userName"] == null ? "none" : "inline")">
                         <a data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" asp-controller="Projects" asp-action="Index">
                             Projects
