+++ Source/FaaS.MVC/wwwroot/.gitignore
@@ -1,0 +1,1 @@
+﻿out/*.out.js
