+++ Source/FaaS.MVC/wwwroot/_references.js
@@ -5,3 +5,11 @@
 /// <reference path="lib/jquery/dist/jquery.js" />
 /// <reference path="lib/jquery-validation/dist/jquery.validate.js" />
 /// <reference path="lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js" />
+/// <reference path="out/dashboard.out.js" />
+/// <reference path="out/form.out.js" />
+/// <reference path="out/login.out.js" />
+/// <reference path="src/dashboard.js" />
+/// <reference path="src/form.js" />
+/// <reference path="src/index.js" />
+/// <reference path="src/login.js" />
+/// <reference path="webpack.config.js" />
