+++ Source/FaaS.MVC/wwwroot/src/dashboard.js
@@ -1,0 +1,1 @@
+﻿
