+++ Source/FaaS.MVC/wwwroot/src/form.js
@@ -1,0 +1,1 @@
+﻿
