+++ Source/FaaS.MVC/wwwroot/src/index.js
@@ -1,0 +1,2 @@
+﻿import React from "react";
+import ReactDOM from "react-dom";
