+++ Source/FaaS.MVC/wwwroot/src/login.js
@@ -1,0 +1,1 @@
+﻿
